pyproject.toml
[tool.poetry]
name = "remote-processor-service"
version = "0.1.0"
description = "A service for hosting search processors extrnal to opensearch"
authors = ["HenryL27 <hmlindeman@yahoo.com>"]
readme = "README.md"
packages = [
    { include = "lib" },
    { include = "service" },
    { include = "gen" },
    { include = "test" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
click = "^8.1.7"
pyyaml = "^6.0.1"
grpcio = "1.60.0"
grpcio-tools = "1.60.0"
cbor2 = "^5.6.0"
jupyter = "^1.0.0"
sycamore-ai = {path = "../sycamore", develop = true}

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"
pytest-mock = "^3.12.0"
docker = "^7.0.0"
opensearch-py = "^2.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
test_cfg = "service.cli:read_cfg"
server = "service.cli:serve"

[tool.pytest.ini_options]
markers = [
    "processor_name: marks integ test with name of processor to test",
]